Avoid pulling out hydrateClient
