+++ packages/sku/src/services/vite/entries/ssr-client.tsx
@@ -1,10 +1,13 @@
 import 'virtual:sku/polyfills';
+import { hydrateRoot } from 'react-dom/client';
+import { createBrowserRouter, matchRoutes, RouterProvider } from 'react-router';
 // Resolved by sku's Vite config plugin to the consumer client / routes entries.
 import * as clientEntry from '__sku_alias__clientEntry';
 import * as routesEntry from '__sku_alias__routesEntry';
+import { SkuProvider } from '#runtime/skuContext';
 import { registerSiteRouteTree } from '#runtime/preloadRoute';
+import { Document } from '../ssr/Document.js';
 import { buildSiteRouteTrees } from '../ssr/buildSiteRouteTrees.js';
-import { hydrateClient } from '../ssr/hydrateClient.js';
 import { readRoutesEntry } from '../ssr/readRoutesEntry.js';
 import { assertSiteName, selectForSite } from '../ssr/selectForSite.js';
 import type { SkuClientEntry } from '../ssr/types.js';
@@ -30,20 +33,53 @@ const hydrate = async () => {
   // `usePreloadRoute` matches against the same tree the router navigates.
   registerSiteRouteTree(siteRoutes);
   const clientContext = window.__SKU_CLIENT_CONTEXT__;
+  onHydrate?.({ clientContext });
 
-  await hydrateClient({
-    site,
-    clientContext,
-    siteRoutes,
-    documentAssets: window.__SKU_DOCUMENT_ASSETS__ ?? {
-      css: [],
-      modulePreloads: [],
-    },
-    onHydrate,
-    getReactContext,
-    getRouterContext,
-    instrumentations,
+  const reactContext = await getReactContext?.({ site, clientContext });
+
+  const lazyMatches = matchRoutes(siteRoutes, window.location)?.filter(
+    ({ route }) => route.lazy,
+  );
+
+  await Promise.all(
+    lazyMatches?.map(async ({ route }) => {
+      const lazy = route.lazy;
+      if (typeof lazy !== 'function') {
+        return;
+      }
+      Object.assign(route, await lazy(), { lazy: undefined });
+    }) ?? [],
+  );
+
+  const router = createBrowserRouter(siteRoutes, {
+    ...(getRouterContext
+      ? {
+          getContext: () =>
+            getRouterContext({ site, clientContext, reactContext }),
+        }
+      : {}),
+    ...(instrumentations === undefined ? {} : { instrumentations }),
   });
+
+  hydrateRoot(
+    document,
+    <Document
+      assets={
+        window.__SKU_DOCUMENT_ASSETS__ ?? {
+          css: [],
+          modulePreloads: [],
+        }
+      }
+    >
+      <SkuProvider
+        site={site}
+        clientContext={clientContext}
+        reactContext={reactContext}
+      >
+        <RouterProvider router={router} />
+      </SkuProvider>
+    </Document>,
+  );
 };
 
 hydrate().catch((error: unknown) => {
